auth and token added

+++ backend/src/routes/deleteListing.js
@@ -1,3 +1,4 @@
+import * as admin from 'firebase-admin';
 import { db } from "../db";
 
 export const deleteListingRoute = {
@@ -5,12 +6,14 @@ export const deleteListingRoute = {
     path: '/api/listings/{id}',
     handler: async (req, h) => {
         const { id } = req.params;
-
+        const token = req.headers.authtoken;
+        const user = await admin.auth().verifyIdToken(token);
+        const userId = user.user_id;
          //! TODO: make sure current user is the owner of listings
 
          await db.query(
-             'DELETE FROM listings WHERE id=?',
-             [id],
+             'DELETE FROM listings WHERE id=? AND user_id=?',
+             [id, userId],
          );
          return { message: 'Success!'}
     }
